kraken: convert to TS
